refactor(throttler): implement global and per-route throttling
Refactors the throttler implementation to use a global configuration with the ability to override settings on a per-route basis.

- Centralises throttler configuration in `src/config/throttler.config.ts`.
- Implements a global `ThrottlerGuard` in `app.module.ts` for unauthenticated routes.
- Uses the `@Throttle()` decorator in `assessor.controller.ts` to apply stricter limits for authenticated routes.
- Extracts environment variable validation to a shared Zod schema in `src/config/env.schema.ts` to ensure consistency between runtime and compile-time configurations.
- Removes redundant module-specific throttler configurations.
