+++ src/config/index.ts
@@ -1,3 +1,3 @@
 export { ConfigModule } from './config.module';
 export { ConfigService } from './config.service';
-export type { Config } from './config.service';
+export type { Config } from './env.schema';
