+++ src/status/status.controller.ts
@@ -1,4 +1,5 @@
-import { Controller, Get, HttpException } from '@nestjs/common';
+import { Controller, Get, HttpException, UseGuards } from '@nestjs/common';
+import { ThrottlerGuard } from '@nestjs/throttler';
 
 import { StatusService, HealthCheckResponse } from './status.service';
